+++ database/models/Product.rb
@@ -78,16 +78,9 @@ class Product < Sequel::Model
     end
   end
 
-  def self.get_statistic(plans)
-    ResultSet.where(plan_id: plans.map(&:id)).group_and_count(:plan_id, :status).map(&:values).group_by do |e|
+  def self.get_statistic(plan_ids)
+    ResultSet.where(plan_id: plan_ids).group_and_count(:plan_id, :status).map(&:values).group_by do |e|
       e[:plan_id]
-    end
-  end
-
-  def self.add_statictic(plans)
-    statistic = get_statistic(plans)
-    plans.map(&:values).map do |plan|
-      plan.merge!(statistic: statistic[plan[:id]] || [])
     end
   end